Replace lua_load with luaL_loadstring
lua_load was giving problems on some machines but not others;
the problem appears to occur on the second return from Lua_SAtt_Reader,
but for the life of me I can't figure out what the Lua interpreter
thinks it is seeing.

Also fixed a bug: each line in the input file gets to be a SAttr,
but when they are reassembled into a Lua script, the newlines were
not being reintroduced. Now they are.

Old version of the code is still present because it's the right
way to do things, if only it worked.